Add test that required attributes are present
